requirements.txt
groq
requests
streamlit
langchain
langchain-community
faiss-cpu
huggingface-hub
sentence-transformers
python-dotenv
numpy
pypdf
pip install -r requirements.txt
